fix: 로그인 돼 있을 때 다시 로그인 안되는 문제 해결

+++ src/main/java/teamproject1/letsdoit/common/config/handler/CustomSimpleUrlAuthenticationFailureHandler.java
@@ -31,8 +31,11 @@ public class CustomSimpleUrlAuthenticationFailureHandler extends SimpleUrlAuthen
                 .queryParam("error", exception.getLocalizedMessage())
                 .build().toUriString();
 
-        customAuthorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
-
-        getRedirectStrategy().sendRedirect(request, response, targetUrl);
+        if (CustomCookie.getCookie(request, "email").isPresent()) {
+            getRedirectStrategy().sendRedirect(request, response, "/home");
+        }else {
+            customAuthorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
+            getRedirectStrategy().sendRedirect(request, response, targetUrl);
+        }
     }
 }
